Update README.md
It is known that each number entered in the Collatz function after some time falls to the value 1 and consequently enters the infinite loop 4-2-1.
The reason for this is the two rules by which the function is realized. If the value is an odd number, the rule 3 * X + 1 applies and if the value is an even number, the rule X / 2.
One is an odd number and is multiplied by three plus 1 which gives 4. Four is an even number and is divisible by two. Two is an even number and is divisible by 2 so that we come to the value 1 again.So far, a check has been made for all values up to 2 ^ 65 and they all ended up in a 4-2-1 loop. So far, no evidence has been found as to why this is happening and whether it applies to all numbers.The algorithm in front of you introduces the desired value into the function and calculates how many divisions and multipliers with adding of number 1 there were   in the process until the number falls into the 4-2-1 loop.It also calculates the sum of subtraction of value  by division and the sum of increase in value  by multiplication  with  addition.It also calculates the difference between the sum of increments and the sum of decrements. This difference is crucial because when it is subtracted from the value entered at the beginning of the process, a value of 1 is obtained, which is the value that each number reaches at the end.
Due to this specific difference,created by the higher number of divisions in relation to the number of multiplications, which is a consequence of adding the number 1 after each multiplication, each number entered in the function will inevitably fall to a value of 1 and this applies to all positive integers.